lots of editing

+++ controllers/cardControllers.js
@@ -3,17 +3,16 @@ const { v4: uuidv4 } = require('uuid');
 const {startPayment, completePayment} = require('../services/payment');
 const cardModel = require('../models/cardModel')
 const userModel = require('../models/userModels')
-const {validateAddCard} = require('../validations/addCardValidations')
+const {validateAddCard, validateFindCard, validateDeleteCard} = require('../validations/addCardValidations')
+const { Op } = require("sequelize");
+const logger = require('../utils/logger')
 
 const startAddCard = async (req,res)=>{
     try{
         const {error} = validateAddCard(req.body)
         if (error !== undefined) {
-            res.status(400).json({
-                status: true,
-                message: error.details[0].message || "Bad request"
-            })
-            return
+            throw new Error (error.details[0].message, 400) 
+            
         }
         const {amount, email }= req.body
 
@@ -23,23 +22,22 @@ const startAddCard = async (req,res)=>{
             }
         });
         if(!CheckIfUserIsRegistered){
-            res.status(400).json({
-                status: false,
-                message: "User is not registered"
-        })
-        return
+            throw new Error ("User not registered", 400)
     }
     
     const initializeCardTransaction = await startPayment(amount, email);
     if (!initializeCardTransaction.data.status){
-    res.status(400).json({
-        status: false,
-        message: 'Invalid transaction'
-    })
+        throw new Error ("Invalid transaction", 500)
     }
-
-    res.json (initializeCardTransaction.data)
-}catch (err) {console.log (err)}
+delete initializeCardTransaction.data.data.access_code
+logger(email,true,"User initiated transaction")
+.then(()=>{res.status(200).json (initializeCardTransaction.data)}
+    )
+    
+}catch (err) {res.json({
+    status: false,
+    message: err.message
+})}
     }
 
 
@@ -48,18 +46,12 @@ const completeAddCard = async (req,res) =>{
     const ref= req.body.reference
     const completeCardTransaction= await completePayment(ref)
     if (!completeCardTransaction.data.status){
-    res.status(400).json({
-        status: false,
-        message: 'Unable to complete transaction'
-    })
+    throw new Error ('Unable to complete card transaction')
 }
 
     const {status,authorization,customer,amount} = completeCardTransaction.data.data
     if (status !== "success"){
-    res.status(400).json({
-        status: false,
-        message: 'Unable to complete transaction'
-    })
+        throw new Error ('Unable to complete card transaction')
         }
 
 
@@ -70,12 +62,7 @@ const completeAddCard = async (req,res) =>{
     }
     });
     if (checkIfCardExist){
-    res.status(400).json({
-        status: false,
-        message: 'Card already exists'
-    })
-    
-    return
+        throw new Error ('Card already exists')
     
 }
     const findUserId= await userModel.findOne({
@@ -85,16 +72,13 @@ const completeAddCard = async (req,res) =>{
                      
                  }
              })
-             if(!findUserId){
-                res.status(400).json({
-                    status: false,
-                    message: 'User not found'
-                })
-                return
+             if(!findUserId.dataValues.user_id){
+                throw new Error ('User not found')
              }
+            
              const amountInNaira = amount/ 100
              const comments = `Wallet funding of ${amountInNaira} was successful`
-await credit(amountInNaira,findUserId.user_id,comments)
+//await credit(amountInNaira,findUserId.dataValues.user_id,comments)
     
 await cardModel.create({
     card_id: uuidv4(),
@@ -118,11 +102,20 @@ await cardModel.create({
     
                 })
 return 
-}catch (err) {console.log (err)}}
+}catch (err) {res.json({
+    status: false,
+    message: err.message
+})}}
+
 
 
 const findCard = async(req,res) => {
+    try{
 const {email} = req.body
+const {error} = validateFindCard(req.body)
+        if (error !== undefined) {
+            throw new Error (error.details[0].message, 400) 
+}
 const cardsFound = await cardModel.findAll({
     
      where: {
@@ -130,14 +123,39 @@ const cardsFound = await cardModel.findAll({
          
      }
  })
- res.status(200).json({
+ if (cardsFound.length === 0){
+    throw new Error ('Card not found')
+ }
+res.status(200).json({
     status : true,
     data: cardsFound
  })
+}catch(e){
+    res.status(500).json({
+        status : false,
+        message : e.message
+    })
+}
 }
 
 const deleteCard = async (req, res) => {
-    const {email,authorization_code} = req.body
+    try{
+    const {error} = validateDeleteCard(req.body)
+    if (error !== undefined) {
+        throw new Error (error.details[0].message, 400) 
+}
+const {email,authorization_code} = req.body
+const cardsFound = await cardModel.findAll({
+    
+    where: {
+        email: email
+        
+    }
+})
+if (cardsFound.length === 0){
+   throw new Error ('Card not found')
+}
+    
     await cardModel.destroy({
         where: {
           [Op.and]: [
@@ -150,7 +168,12 @@ const deleteCard = async (req, res) => {
         status : true,
         message: "card deleted"
      })
-     
+    }catch(e){
+        res.status(500).json({
+            status : false,
+            message : e.message
+        })
+    }
    }
 
 module.exports = {startAddCard, completeAddCard, findCard,deleteCard}
